Move SnackBar and VoterInput to their own files
